Travis tweaks
- No longer wait 60 seconds for app startup in `before_script`, now we check if the application is up right before we need it. If it's not up we wait for it to be up. This allows the checks to start running while the application is starting up in the background.
- Regular e2e tests do not run on a Travis `tag` build anymore, by the time we tag a commit, these tests will already have run and succeeded.
- Production e2e tests only run on release branches
- Docker builds only run on release branches

+++ src/test/bash/run-prod-e2e.sh
@@ -10,23 +10,23 @@
 
 set -ev
 
-sed -i "s|new plugin.BaseHrefWebpackPlugin({ baseHref: '/' })|new plugin.BaseHrefWebpackPlugin({ baseHref: '/managementportal/' })|" webpack/webpack.dev.js
-sed -i "s|baseUrl: 'http://localhost:8080/',|baseUrl: 'http://localhost:8080/managementportal/',|" src/test/javascript/protractor.conf.js
-sed -i "s|contexts: prod|contexts: dev|" src/main/resources/config/application-prod.yml # set liquibase context to dev so it loads demo data
-./gradlew bootRepackage -Pprod buildDocker -x test
-docker-compose -f src/main/docker/app.yml up -d # spin up production mode application
-# wait for app to start up
-echo "Waiting for application startup"
-until curl -s http://localhost:8080/managementportal/ > /dev/null
-do
-  echo -n "." # waiting
-  sleep 2
-done
-echo ""
-echo "Application is up"
-docker-compose -f src/main/docker/app.yml logs # show output of app startup
-yarn e2e # run e2e tests against production mode
-docker-compose -f src/main/docker/app.yml down -v # clean up containers and volumes
-git checkout src/test/javascript/protractor.conf.js
-git checkout webpack/webpack.prod.js
-git checkout src/main/resources/config/application-prod.yml
+# only run on the release branch's push and pull_request events
+if [[ $TRAVIS_BRANCH == release-* ]] || [[ $TRAVIS_PULL_REQUEST_BRANCH == release-* ]]
+then
+  echo "Running production e2e tests"
+  sed -i "s|new plugin.BaseHrefWebpackPlugin({ baseHref: '/' })|new plugin.BaseHrefWebpackPlugin({ baseHref: '/managementportal/' })|" webpack/webpack.dev.js
+  sed -i "s|baseUrl: 'http://localhost:8080/',|baseUrl: 'http://localhost:8080/managementportal/',|" src/test/javascript/protractor.conf.js
+  sed -i "s|contexts: prod|contexts: dev|" src/main/resources/config/application-prod.yml # set liquibase context to dev so it loads demo data
+  ./gradlew bootRepackage -Pprod buildDocker -x test
+  docker-compose -f src/main/docker/app.yml up -d # spin up production mode application
+  # wait for app to be up
+  ./../../../util/wait-for-app.sh
+  docker-compose -f src/main/docker/app.yml logs # show output of app startup
+  yarn e2e # run e2e tests against production mode
+  docker-compose -f src/main/docker/app.yml down -v # clean up containers and volumes
+  git checkout src/test/javascript/protractor.conf.js
+  git checkout webpack/webpack.prod.js
+  git checkout src/main/resources/config/application-prod.yml
+else
+  echo "Skipping production e2e tests"
+fi
